+++ src/gpu_wrapper.h
@@ -47,7 +47,7 @@ class GPUWrapper {
     struct descentgraph* gpu_init_descentgraph();
     curandState* gpu_init_random_curand();
     tinymt32_status_t* gpu_init_random_tinymt();
-    double* gpu_init_lodscores();
+    float* gpu_init_lodscores();
     struct founderallelegraph* gpu_init_founderallelegraph();
     
     void copy_to_gpu(DescentGraph& dg);
